backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
opencv-python-headless==4.8.1.78
torch==2.1.1
torchvision==0.16.1
librosa==0.10.1
numpy==1.26.2
Pillow==10.1.0
requests==2.31.0
gdown==5.1.0
werkzeug==3.0.1
python-dotenv==1.0.0
soundfile==0.12.1
yt-dlp
